+++ BackEnd/src/main/java/com/sansantek/sansanmulmul/mountain/service/MountainService.java
@@ -2,9 +2,11 @@ package com.sansantek.sansanmulmul.mountain.service;
 
 
 import com.sansantek.sansanmulmul.mountain.domain.Mountain;
+import com.sansantek.sansanmulmul.mountain.domain.spot.MountainSpot;
 import com.sansantek.sansanmulmul.mountain.repository.MountainRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 
 import java.util.List;
 
@@ -14,11 +16,12 @@ public class MountainService {
     @Autowired
     private MountainRepository mountainRepository;
 
+
     public List<Mountain> getAllMountains(){
         return mountainRepository.findAll();
     }
 
-    public Mountain getMountainDetail(Long mountain_id){
+    public Mountain getMountainDetail(int mountain_id){
         return mountainRepository.findById(mountain_id).orElse(null);
     }
